Run digest cycle on graph's zoom and pan events
Rationale: we want to add "fit" and "zoom to central object" buttons
that should be visible when graph is not fitted or is not zoomed to
the center respectively. It's simple to write such conditions for
buttons but they need to be recalculated as the graph is zoomed or
panned.